Allow to overwrite app_user variable
We have some older projects at Simplificator where we use different names for our app_user than the common standard today. With the previous solution, we were unable to also distribute the keys to those old users.

By converting the list of keys to a JSON, Ansible is able to use variables as keys and we can overwrite the app_user variable in the few host vars.