Add user to groups rather than overwriting all its groups
When a provisioning failed, it could happen that the app_user was removed from groups which would be added later, because "append" in the user module is per default "no". By changing this to "yes", the user will be added to the mentioned group and will preserve all its other groups.